Improve :poll modularization
