+++ feature/poll/impl/src/main/java/com/ivy/poll/impl/PollScreen.kt
@@ -1,4 +1,4 @@
-package com.ivy.poll
+package com.ivy.poll.impl
 
 import androidx.compose.runtime.Composable
 import com.ivy.navigation.screenScopedViewModel
