+++ feature/poll/impl/src/main/java/com/ivy/poll/impl/PollViewModel.kt
@@ -1,4 +1,4 @@
-package com.ivy.poll
+package com.ivy.poll.impl
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Stable
