+++ feature/poll/public/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+  id("ivy.feature")
+}
+
+android {
+  namespace = "com.ivy.poll"
+}
+
+dependencies {
+}
