+++ settings.gradle.kts
@@ -34,7 +34,8 @@ include(":feature:main")
 include(":feature:onboarding")
 include(":feature:piechart")
 include(":feature:planned-payments")
-include(":feature:poll")
+include(":feature:poll:impl")
+include(":feature:poll:public")
 include(":feature:releases")
 include(":feature:reports")
 include(":feature:search")
